Move runtime and module into module package
A bit messy doing this in a different commit, but after the rename the tunime package makes even less sense to remain distinct.
